+++ Visual Studio/Misc/LeetCode OJ/268. Missing Number/Solution.h
@@ -5,6 +5,6 @@ class Solution
 public:
     int missingNumber(const vector<int> &nums)
     {
-        return nums.size() * (nums.size() + 1) / 2 - accumulate(nums.cbegin(), nums.cend(), 0);
+        return static_cast<int>(nums.size() * (nums.size() + 1) / 2 - accumulate(nums.cbegin(), nums.cend(), 0));
     }
 };
